requirements.txt
taichi>=1.6.0
matplotlib
numpy
pytorch_msssim
dataclass-wizard
pillow
pyyaml
pandas[parquet]>=2.0.0
scipy
argparse
tensorboard
plyfile